Using url ,screenshot and manually generating is done
