+++ static/js/preview.js
@@ -1,0 +1,14 @@
+// Preview the image before uploading
+function previewImage(event) {
+    var reader = new FileReader();
+    var imageField = document.getElementById("preview");
+
+    reader.onload = function() {
+        if (reader.readyState == 2) {
+            imageField.src = reader.result;
+            imageField.style.display = 'block';  // Show the image preview
+        }
+    };
+
+    reader.readAsDataURL(event.target.files[0]);
+}
