+++ templates/display_image.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
+    <title>Image Generated</title>
+</head>
+<body>
+    <h1>Generated Image</h1>
+    <img src="{{ url_for('uploaded_file', filename=filename) }}" alt="Generated Image"><br><br>
+    <a href="/">Go Back</a>
+</body>
+</html>
